+++ relay-server/src/actors/store.rs
@@ -50,10 +50,38 @@ pub enum StoreError {
     NoEventId,
 }
 
+type Producer = Arc<ThreadedProducer>;
+
+struct Producers {
+    events: Producer,
+    attachments: Producer,
+    transactions: Producer,
+    sessions: Producer,
+    metrics: Producer,
+}
+
+impl Producers {
+    /// Get a producer by KafkaTopic value
+    pub fn get(&self, kafka_topic: KafkaTopic) -> Option<&Producer> {
+        match kafka_topic {
+            KafkaTopic::Attachments => Some(&self.attachments),
+            KafkaTopic::Events => Some(&self.events),
+            KafkaTopic::Transactions => Some(&self.transactions),
+            KafkaTopic::Outcomes => {
+                // should be unreachable
+                relay_log::error!("attempted to send data to outcomes topic from store forwarder. there is another actor for that.");
+                None
+            }
+            KafkaTopic::Sessions => Some(&self.sessions),
+            KafkaTopic::Metrics => Some(&self.metrics),
+        }
+    }
+}
+
 /// Actor for publishing events to Sentry through kafka topics.
 pub struct StoreForwarder {
     config: Arc<Config>,
-    producer: Arc<ThreadedProducer>,
+    producers: Producers,
 }
 
 fn make_distinct_id(s: &str) -> Uuid {
@@ -61,21 +89,50 @@ fn make_distinct_id(s: &str) -> Uuid {
         .unwrap_or_else(|_| Uuid::new_v5(&NAMESPACE_DID, s.as_bytes()))
 }
 
+/// Temporary map used to deduplicate kafka producers
+type ReusedProducersMap<'a> = BTreeMap<Option<&'a str>, Producer>;
+
+fn make_producer<'a>(
+    config: &'a Config,
+    reused_producers: &mut ReusedProducersMap<'a>,
+    kafka_topic: KafkaTopic,
+) -> Result<Producer, ServerError> {
+    let (config_name, kafka_config) = config
+        .kafka_config(kafka_topic)
+        .context(ServerErrorKind::KafkaError)?;
+
+    if let Some(producer) = reused_producers.get(&config_name) {
+        return Ok(Arc::clone(producer));
+    }
+
+    let mut client_config = ClientConfig::new();
+
+    for config_p in kafka_config {
+        client_config.set(config_p.name.as_str(), config_p.value.as_str());
+    }
+
+    let producer = Arc::new(
+        client_config
+            .create_with_context(CaptureErrorContext)
+            .context(ServerErrorKind::KafkaError)?,
+    );
+
+    reused_producers.insert(config_name, Arc::clone(&producer));
+    Ok(producer)
+}
+
 impl StoreForwarder {
     pub fn create(config: Arc<Config>) -> Result<Self, ServerError> {
-        let mut client_config = ClientConfig::new();
-        for config_p in config.kafka_config() {
-            client_config.set(config_p.name.as_str(), config_p.value.as_str());
-        }
+        let mut reused_producers = BTreeMap::new();
+        let producers = Producers {
+            attachments: make_producer(&*config, &mut reused_producers, KafkaTopic::Attachments)?,
+            events: make_producer(&*config, &mut reused_producers, KafkaTopic::Events)?,
+            transactions: make_producer(&*config, &mut reused_producers, KafkaTopic::Transactions)?,
+            sessions: make_producer(&*config, &mut reused_producers, KafkaTopic::Sessions)?,
+            metrics: make_producer(&*config, &mut reused_producers, KafkaTopic::Metrics)?,
+        };
 
-        let producer = client_config
-            .create_with_context(CaptureErrorContext)
-            .context(ServerErrorKind::KafkaError)?;
-
-        Ok(Self {
-            config,
-            producer: Arc::new(producer),
-        })
+        Ok(Self { config, producers })
     }
 
     fn produce(&self, topic: KafkaTopic, message: KafkaMessage) -> Result<(), StoreError> {
@@ -86,10 +143,13 @@ impl StoreForwarder {
             .key(&key)
             .payload(&serialized);
 
-        match self.producer.send(record) {
-            Ok(_) => Ok(()),
-            Err((kafka_error, _message)) => Err(StoreError::SendFailed(kafka_error)),
+        if let Some(producer) = self.producers.get(topic) {
+            producer
+                .send(record)
+                .map_err(|(kafka_error, _message)| StoreError::SendFailed(kafka_error))?;
         }
+
+        Ok(())
     }
 
     fn produce_attachment_chunks(