+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
-**Features**
+**Features**:
 
 - Add bucket width to bucket protocol. ([#1103](https://github.com/getsentry/relay/pull/1103))
-
+- Support multiple kafka cluster configurations ([#1101](https://github.com/getsentry/relay/pull/1101))
 
 ## 21.10.0